// lib/api/generated/models/eventWithdrawal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { WithdrawalStatus } from './withdrawalStatus';

export interface EventWithdrawal {
  at: string;
  athlete_id: string;
  reason: string;
  status: WithdrawalStatus;
  /** @nullable */
  user_id?: string | null;
}