// lib/api/generated/models/publicCalendarEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export interface PublicCalendarEvent {
  /** @nullable */
  cancellation_note?: string | null;
  date: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  end_date?: string | null;
  event_type: string;
  id: string;
  /** @nullable */
  location?: string | null;
  status: string;
  /** @nullable */
  time?: string | null;
  title: string;
}