// lib/api/generated/models/trainingScheduleDefaults.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export interface TrainingScheduleDefaults {
  /** @minimum 0 */
  attendance_buffer_minutes: number;
  end_time: string;
  location: string;
  time: string;
  title: string;
  /**
     * ISO weekdays 1=Mon … 7=Sun
     * @items.minimum 0
     */
  weekdays: number[];
}